requirements.txt
pandas
numpy
pyyaml
typer[all]